+++ src/main/java/org/folio/search/controller/SearchConsortiumController.java
@@ -11,6 +11,7 @@ import org.folio.search.domain.dto.BatchIdsDto;
 import org.folio.search.domain.dto.ConsortiumCampusCollection;
 import org.folio.search.domain.dto.ConsortiumHolding;
 import org.folio.search.domain.dto.ConsortiumHoldingCollection;
+import org.folio.search.domain.dto.ConsortiumInstitutionCollection;
 import org.folio.search.domain.dto.ConsortiumItem;
 import org.folio.search.domain.dto.ConsortiumItemCollection;
 import org.folio.search.domain.dto.ConsortiumLibraryCollection;
@@ -25,6 +26,7 @@ import org.folio.search.rest.resource.SearchConsortiumApi;
 import org.folio.search.service.consortium.ConsortiumCampusService;
 import org.folio.search.service.consortium.ConsortiumInstanceSearchService;
 import org.folio.search.service.consortium.ConsortiumInstanceService;
+import org.folio.search.service.consortium.ConsortiumInstitutionService;
 import org.folio.search.service.consortium.ConsortiumLibraryService;
 import org.folio.search.service.consortium.ConsortiumLocationService;
 import org.folio.search.service.consortium.ConsortiumTenantService;
@@ -50,6 +52,7 @@ public class SearchConsortiumController implements SearchConsortiumApi {
   private final ConsortiumInstanceSearchService searchService;
   private final ConsortiumCampusService campusService;
   private final ConsortiumLibraryService libraryService;
+  private final ConsortiumInstitutionService institutionService;
 
   @Override
   public ResponseEntity<ConsortiumHoldingCollection> getConsortiumHoldings(String tenantHeader, String instanceId,
@@ -128,6 +131,21 @@ public class SearchConsortiumController implements SearchConsortiumApi {
     return ResponseEntity.ok(new
       ConsortiumLibraryCollection()
       .libraries(result.getRecords())
+      .totalRecords(result.getTotalRecords()));
+  }
+
+  @Override
+  public ResponseEntity<ConsortiumInstitutionCollection> getConsortiumInstitutions(String tenantHeader,
+                                                                                   String tenantId,
+                                                                                   Integer limit,
+                                                                                   Integer offset,
+                                                                                   String sortBy,
+                                                                                   SortOrder sortOrder) {
+    var result = institutionService.fetchInstitutions(tenantHeader, tenantId, limit, offset, sortBy, sortOrder);
+
+    return ResponseEntity.ok(new
+      ConsortiumInstitutionCollection()
+      .institutions(result.getRecords())
       .totalRecords(result.getTotalRecords()));
   }